test: enhance test coverage for toast
